Refactor SBML processor
