+++ mira/sources/sbml/processor.py
@@ -15,7 +15,6 @@ from typing import Dict, Iterable, List, Mapping, Optional, Tuple
 import bioregistry
 import curies
 import sympy
-from libsbml import SBMLReader
 from lxml import etree
 from tqdm import tqdm
 
@@ -34,14 +33,6 @@ from mira.metamodel import (
 )
 from mira.resources import get_resource_file
 
-__all__ = [
-    "template_model_from_sbml_file",
-    "template_model_from_sbml_file_obj",
-    "template_model_from_sbml_string",
-    "template_model_from_sbml_model",
-    "sbml_model_from_file"
-]
-
 
 class TqdmLoggingHandler(logging.Handler):
     def __init__(self, level=logging.NOTSET):
@@ -54,6 +45,7 @@ class TqdmLoggingHandler(logging.Handler):
             self.flush()
         except Exception:
             self.handleError(record)
+
 
 logger = logging.getLogger(__name__)
 logger.addHandler(TqdmLoggingHandler())
@@ -105,274 +97,233 @@ class Converter:
 converter = Converter()
 
 
-def template_model_from_sbml_file(
-        file_path,
-        *,
-        model_id: Optional[str] = None,
-        reporter_ids: Optional[Iterable[str]] = None,
-) -> TemplateModel:
-    """Extract a MIRA template model from a file containing SBML XML."""
-    sbml_model = sbml_model_from_file(file_path)
-    return template_model_from_sbml_model(sbml_model, model_id=model_id,
-                                          reporter_ids=reporter_ids)
+class SbmlProcessor:
+    """Traverse an SBML document generated by libSBML and extract a MIRA
+    Template model."""
 
+    def __init__(self, sbml_model, model_id=None, reporter_ids=None):
+        self.sbml_model = sbml_model
+        self.model_id = model_id
+        self.reporter_ids = reporter_ids
 
-def template_model_from_sbml_file_obj(
-    file,
-    *,
-    model_id: Optional[str] = None,
-    reporter_ids: Optional[Iterable[str]] = None,
-) -> TemplateModel:
-    """Extract a MIRA template model from a file object containing SBML XML."""
-    return template_model_from_sbml_string(
-        file.read().decode("utf-8"), model_id=model_id, reporter_ids=reporter_ids
-    )
+    def extract_model(self):
+        if self.model_id is None:
+            model_id = get_model_id(self.sbml_model)
+        reporter_ids = set(self.reporter_ids or [])
+        concepts = _extract_concepts(self.sbml_model, model_id=model_id)
 
+        def _lookup_concepts_filtered(species_ids) -> List[Concept]:
+            return [
+                concepts[species_id] for species_id in species_ids
+                if species_id not in reporter_ids and 'cumulative' not in species_id
+            ]
 
-def sbml_model_from_file(fname):
-    """Return an SBML model object from an SBML file."""
-    with open(fname, 'rb') as fh:
-        sbml_string = fh.read().decode('utf-8')
-    sbml_document = SBMLReader().readSBMLFromString(sbml_string)
-    return sbml_document.getModel()
+        # Iterate thorugh all reactions and piecewise convert to templates
+        templates: List[Template] = []
+        # see docs on reactions
+        # https://sbml.org/software/libsbml/5.18.0/docs/formatted/python-api/
+        # classlibsbml_1_1_reaction.html
+        all_species = {species.id for species in self.sbml_model.species}
+        all_parameters = {parameter.id: parameter.value
+                          for parameter in self.sbml_model.parameters}
+        parameter_symbols = {parameter.id: sympy.Symbol(parameter.id)
+                             for parameter in self.sbml_model.parameters}
+        compartment_symbols = {compartment.id: sympy.Symbol(compartment.id)
+                               for compartment in self.sbml_model.compartments}
+        # Add compartment volumes as parameters
+        for compartment in self.sbml_model.compartments:
+            all_parameters[compartment.id] = compartment.volume
+        if 'lambda' in all_parameters:
+            all_parameters['XXlambdaXX'] = all_parameters.pop('lambda')
+        if 'lambda' in parameter_symbols:
+            parameter_symbols.pop('lambda')
+            parameter_symbols['XXlambdaXX'] = sympy.Symbol('XXlambdaXX')
 
+        # Handle custom function definitions in the model
+        function_lambdas = {}
+        for fun_def in self.sbml_model.function_definitions:
+            args = [fun_def.getArgument(i).getName()
+                    for i in range(fun_def.getNumArguments())]
+            arg_symbols = {arg: sympy.Symbol(arg) for arg in args}
+            if 'lambda' in arg_symbols:
+                arg_symbols.pop('lambda')
+                arg_symbols['XXlambdaXX'] = sympy.Symbol('XXlambdaXX')
 
-def template_model_from_sbml_string(
-    s: str,
-    *,
-    model_id: Optional[str] = None,
-    reporter_ids: Optional[Iterable[str]] = None,
-) -> TemplateModel:
-    """Extract a MIRA template model from a string representing SBML XML."""
-    sbml_document = SBMLReader().readSBMLFromString(s)
-    return template_model_from_sbml_model(
-        sbml_document.getModel(), model_id=model_id, reporter_ids=reporter_ids
-    )
-
-
-def template_model_from_sbml_model(
-    sbml_model,
-    *,
-    model_id: Optional[str] = None,
-    reporter_ids: Optional[Iterable[str]] = None,
-) -> TemplateModel:
-    """Traverse an SBML document generated by libSBML and extract a MIRA Template model."""
-    if model_id is None:
-        model_id = get_model_id(sbml_model)
-    reporter_ids = set(reporter_ids or [])
-    concepts = _extract_concepts(sbml_model, model_id=model_id)
-
-    def _lookup_concepts_filtered(species_ids) -> List[Concept]:
-        return [
-            concepts[species_id] for species_id in species_ids
-            if species_id not in reporter_ids and 'cumulative' not in species_id
-        ]
-
-    # Iterate thorugh all reactions and piecewise convert to templates
-    templates: List[Template] = []
-    # see docs on reactions
-    # https://sbml.org/software/libsbml/5.18.0/docs/formatted/python-api/classlibsbml_1_1_reaction.html
-    all_species = {species.id for species in sbml_model.species}
-    all_parameters = {parameter.id: parameter.value
-                      for parameter in sbml_model.parameters}
-    parameter_symbols = {parameter.id: sympy.Symbol(parameter.id)
-                         for parameter in sbml_model.parameters}
-    compartment_symbols = {compartment.id: sympy.Symbol(compartment.id)
-                           for compartment in sbml_model.compartments}
-    # Add compartment volumes as parameters
-    for compartment in sbml_model.compartments:
-        all_parameters[compartment.id] = compartment.volume
-    if 'lambda' in all_parameters:
-        all_parameters['XXlambdaXX'] = all_parameters.pop('lambda')
-    if 'lambda' in parameter_symbols:
-        parameter_symbols.pop('lambda')
-        parameter_symbols['XXlambdaXX'] = sympy.Symbol('XXlambdaXX')
-
-    # Handle custom function definitions in the model
-    function_lambdas = {}
-    for fun_def in sbml_model.function_definitions:
-        args = [fun_def.getArgument(i).getName()
-                for i in range(fun_def.getNumArguments())]
-        arg_symbols = {arg: sympy.Symbol(arg) for arg in args}
-        if 'lambda' in arg_symbols:
-            arg_symbols.pop('lambda')
-            arg_symbols['XXlambdaXX'] = sympy.Symbol('XXlambdaXX')
-
-        signature = tuple(arg_symbols.values())
-        formula_str = get_formula_str(fun_def.getBody())
-        if isinstance(formula_str, float) and math.isnan(formula_str):
-            continue
-        formula = sympy.parse_expr(formula_str, local_dict=arg_symbols)
-        lmbd = sympy.Lambda(signature, formula)
-        function_lambdas[fun_def.id] = lmbd
-
-    # In formulas, the species ID appears instead of the species name
-    # and so we have to map these to symbols corresponding to the species name
-    species_id_map = {
-        species.id: sympy.Symbol(species.name)
-        for species in sbml_model.species
-    }
-
-    all_locals = {k: v for k, v in (list(parameter_symbols.items()) +
-                                    list(compartment_symbols.items()) +
-                                    list(function_lambdas.items()) +
-                                    list(species_id_map.items()))}
-
-    # Handle custom assignment rules in the model
-    assignment_rules = {}
-    for rule in sbml_model.rules:
-        rule_expr = parse_assignment_rule(rule.formula, all_locals)
-        if rule_expr:
-            assignment_rules[rule.id] = rule_expr
-
-    for reaction in sbml_model.reactions:
-        modifier_species = [species.species for species in reaction.modifiers]
-        reactant_species = [species.species for species in reaction.reactants]
-        product_species = [species.species for species in reaction.products]
-
-        rate_law = reaction.getKineticLaw()
-
-        def clean_formula(f):
-            return f.replace('lambda', 'XXlambdaXX')
-
-        rate_expr = sympy.parse_expr(clean_formula(rate_law.formula),
-                                     local_dict=all_locals)
-        # At this point we need to make sure we substitute the assignments
-        rate_expr = rate_expr.subs(assignment_rules)
-        for comp, comp_symbol in compartment_symbols.items():
-            # We want to handle the special case where the compartment is
-            # just a constant 1.0 and so we can just remove it from the
-            # rate expression but that requires some special handling otherwise
-            # and explicit 1.0 will be carried around in the rate expression
-            comp_one = False
-            if comp_symbol in rate_expr.free_symbols:
-                if rate_expr not in rate_expr.diff(comp_symbol).free_symbols:
-                    if all_parameters[comp] == 1.0:
-                        comp_one = True
-                        rate_expr /= comp_symbol
-            if not comp_one:
-                rate_expr = rate_expr.subs(comp_symbol, all_parameters[comp])
-
-        rate_law_variables = variables_from_sympy_expr(rate_expr)
-
-        # Implicit modifiers appear in the rate law but are not reactants and
-        # aren't listed explicitly as modifiers. They have to be proper species
-        # though (since the rate law also contains parameters).
-        implicit_modifiers = ((set(rate_law_variables) & all_species)
-                              - (set(reactant_species) | set(modifier_species)))
-        # We extend modifiers with implicit ones
-        modifier_species += sorted(implicit_modifiers)
-
-        modifiers = _lookup_concepts_filtered(modifier_species)
-        reactants = _lookup_concepts_filtered(reactant_species)
-        products = _lookup_concepts_filtered(product_species)
-
-        # check if reaction is reversible (i.e., reversible=False in the attributes),
-        # then add a backwards conversion.
-        if len(reactants) == 1 and len(products) == 1:
-            if reactants[0].name and reactants[0] == products[0]:
-                logger.debug(f"[{model_id} reaction:{reaction.id}]")
-                logger.debug(f"Same reactant and product: {reactants[0]}")
-                logger.debug(f"Modifiers: {modifiers}")
+            signature = tuple(arg_symbols.values())
+            formula_str = get_formula_str(fun_def.getBody())
+            if isinstance(formula_str, float) and math.isnan(formula_str):
                 continue
-            if len(modifiers) == 0:
-                templates.append(
-                    NaturalConversion(
-                        subject=reactants[0],
-                        outcome=products[0],
-                        rate_law=rate_expr,
-                    )
-                )
-            elif len(modifiers) == 1:
-                templates.append(
-                    ControlledConversion(
-                        subject=reactants[0],
-                        outcome=products[0],
-                        controller=modifiers[0],
-                        rate_law=rate_expr,
-                    )
-                )
-            else:
-                # TODO reconsider adding different template that groups multiple controllers
-                """
-                could be the case that there's a linear combination of things that are independent
-                - this could mean you could create multiple conversions
+            formula = sympy.parse_expr(formula_str, local_dict=arg_symbols)
+            lmbd = sympy.Lambda(signature, formula)
+            function_lambdas[fun_def.id] = lmbd
 
-                but, they can be dependent too, then harder to break up
-                """
-                templates.append(
-                    GroupedControlledConversion(
-                        subject=reactants[0],
-                        outcome=products[0],
-                        controllers=modifiers,
-                        rate_law=rate_expr,
-                    )
-                )
-        elif not reactants and not products:
-            logger.debug(f"[{model_id} reaction:{reaction.id}] missing reactants and products")
-            continue
-        elif products and not reactants:
-            if len(products) == 1:
-                if len(modifiers) > 1:
-                    templates.append(
-                        GroupedControlledProduction(
-                            controllers=modifiers,
-                            outcome=products[0],
-                            rate_law=rate_expr,
-                        )
-                    )
-                # todo: handle len(modifiers) == 1, e.g. ControlledProduction?
-                elif len(modifiers) == 1:
-                    logger.debug("Can not yet handle single controller production")
+        # In formulas, the species ID appears instead of the species name
+        # and so we have to map these to symbols corresponding to the species name
+        species_id_map = {
+            species.id: sympy.Symbol(species.name)
+            for species in self.sbml_model.species
+        }
+
+        all_locals = {k: v for k, v in (list(parameter_symbols.items()) +
+                                        list(compartment_symbols.items()) +
+                                        list(function_lambdas.items()) +
+                                        list(species_id_map.items()))}
+
+        # Handle custom assignment rules in the model
+        assignment_rules = {}
+        for rule in self.sbml_model.rules:
+            rule_expr = parse_assignment_rule(rule.formula, all_locals)
+            if rule_expr:
+                assignment_rules[rule.id] = rule_expr
+
+        for reaction in self.sbml_model.reactions:
+            modifier_species = [species.species for species in reaction.modifiers]
+            reactant_species = [species.species for species in reaction.reactants]
+            product_species = [species.species for species in reaction.products]
+
+            rate_law = reaction.getKineticLaw()
+
+            def clean_formula(f):
+                return f.replace('lambda', 'XXlambdaXX')
+
+            rate_expr = sympy.parse_expr(clean_formula(rate_law.formula),
+                                         local_dict=all_locals)
+            # At this point we need to make sure we substitute the assignments
+            rate_expr = rate_expr.subs(assignment_rules)
+            for comp, comp_symbol in compartment_symbols.items():
+                # We want to handle the special case where the compartment is
+                # just a constant 1.0 and so we can just remove it from the
+                # rate expression but that requires some special handling otherwise
+                # and explicit 1.0 will be carried around in the rate expression
+                comp_one = False
+                if comp_symbol in rate_expr.free_symbols:
+                    if rate_expr not in rate_expr.diff(comp_symbol).free_symbols:
+                        if all_parameters[comp] == 1.0:
+                            comp_one = True
+                            rate_expr /= comp_symbol
+                if not comp_one:
+                    rate_expr = rate_expr.subs(comp_symbol, all_parameters[comp])
+
+            rate_law_variables = variables_from_sympy_expr(rate_expr)
+
+            # Implicit modifiers appear in the rate law but are not reactants and
+            # aren't listed explicitly as modifiers. They have to be proper species
+            # though (since the rate law also contains parameters).
+            implicit_modifiers = ((set(rate_law_variables) & all_species)
+                                  - (set(reactant_species) | set(modifier_species)))
+            # We extend modifiers with implicit ones
+            modifier_species += sorted(implicit_modifiers)
+
+            modifiers = _lookup_concepts_filtered(modifier_species)
+            reactants = _lookup_concepts_filtered(reactant_species)
+            products = _lookup_concepts_filtered(product_species)
+
+            # check if reaction is reversible (i.e., reversible=False in the attributes),
+            # then add a backwards conversion.
+            if len(reactants) == 1 and len(products) == 1:
+                if reactants[0].name and reactants[0] == products[0]:
+                    logger.debug(f"[{model_id} reaction:{reaction.id}]")
+                    logger.debug(f"Same reactant and product: {reactants[0]}")
+                    logger.debug(f"Modifiers: {modifiers}")
                     continue
-                else:
+                if len(modifiers) == 0:
                     templates.append(
-                        NaturalProduction(
+                        NaturalConversion(
+                            subject=reactants[0],
                             outcome=products[0],
                             rate_law=rate_expr,
                         )
                     )
-            else:
-                logger.debug("can not yet handle multiple outcome natural production")
-                continue
-        elif reactants and not products:
-            if len(reactants) == 1:
-                templates.append(
-                    NaturalDegradation(
-                        subject=reactants[0],
-                        rate_law=rate_expr,
+                elif len(modifiers) == 1:
+                    templates.append(
+                        ControlledConversion(
+                            subject=reactants[0],
+                            outcome=products[0],
+                            controller=modifiers[0],
+                            rate_law=rate_expr,
+                        )
                     )
-                )
-            else:
-                logger.debug("can not yet handle multiple outcome natural degradation")
+                else:
+                    # TODO reconsider adding different template that groups multiple controllers
+                    """
+                    could be the case that there's a linear combination of things that are independent
+                    - this could mean you could create multiple conversions
+
+                    but, they can be dependent too, then harder to break up
+                    """
+                    templates.append(
+                        GroupedControlledConversion(
+                            subject=reactants[0],
+                            outcome=products[0],
+                            controllers=modifiers,
+                            rate_law=rate_expr,
+                        )
+                    )
+            elif not reactants and not products:
+                logger.debug(f"[{model_id} reaction:{reaction.id}] missing reactants and products")
                 continue
-        else:
-            logger.debug(
-                f"[{model_id} reaction:{reaction.id}] skipping reaction with multiple inputs/outputs"
+            elif products and not reactants:
+                if len(products) == 1:
+                    if len(modifiers) > 1:
+                        templates.append(
+                            GroupedControlledProduction(
+                                controllers=modifiers,
+                                outcome=products[0],
+                                rate_law=rate_expr,
+                            )
+                        )
+                    # todo: handle len(modifiers) == 1, e.g. ControlledProduction?
+                    elif len(modifiers) == 1:
+                        logger.debug("Can not yet handle single controller production")
+                        continue
+                    else:
+                        templates.append(
+                            NaturalProduction(
+                                outcome=products[0],
+                                rate_law=rate_expr,
+                            )
+                        )
+                else:
+                    logger.debug("can not yet handle multiple outcome natural production")
+                    continue
+            elif reactants and not products:
+                if len(reactants) == 1:
+                    templates.append(
+                        NaturalDegradation(
+                            subject=reactants[0],
+                            rate_law=rate_expr,
+                        )
+                    )
+                else:
+                    logger.debug("can not yet handle multiple outcome natural degradation")
+                    continue
+            else:
+                logger.debug(
+                    f"[{model_id} reaction:{reaction.id}] skipping reaction with multiple inputs/outputs"
+                )
+                for i, inp in enumerate(reactants):
+                    logger.debug(f"reactant {i}: {inp!r}")
+                for i, inp in enumerate(products):
+                    logger.debug(f"products {i}: {inp!r}")
+                logger.debug("")
+                continue
+
+        # Gather species-level initial conditions
+        initials = {}
+        for species in self.sbml_model.species:
+            initials[species.name] = Initial(
+                concept=concepts[species.getId()],
+                value=species.initial_concentration,
             )
-            for i, inp in enumerate(reactants):
-                logger.debug(f"reactant {i}: {inp!r}")
-            for i, inp in enumerate(products):
-                logger.debug(f"products {i}: {inp!r}")
-            logger.debug("")
-            continue
 
-    # Gather species-level initial conditions
-    initials = {}
-    for species in sbml_model.species:
-        initials[species.name] = Initial(
-            concept=concepts[species.getId()],
-            value=species.initial_concentration,
-        )
-
-    param_objs = {k: Parameter(name=k, value=v)
-                  for k, v in all_parameters.items()}
-    template_model = TemplateModel(templates=templates,
-                                   parameters=param_objs,
-                                   initials=initials)
-    # Replace constant concepts by their initial value
-    template_model = replace_constant_concepts(template_model)
-    return template_model
+        param_objs = {k: Parameter(name=k, value=v)
+                      for k, v in all_parameters.items()}
+        template_model = TemplateModel(templates=templates,
+                                       parameters=param_objs,
+                                       initials=initials)
+        # Replace constant concepts by their initial value
+        template_model = replace_constant_concepts(template_model)
+        return template_model
 
 
 def get_model_id(sbml_model):
